[Code] fix warning

+++ raygui_styles/style_terminal.h
@@ -15,22 +15,22 @@
 
 // Custom style name: terminal
 static const GuiStyleProp terminalStyleProps[TERMINAL_STYLE_PROPS_COUNT] = {
-    { 0, 0, 0x1c8d00ff },    // DEFAULT_BORDER_COLOR_NORMAL 
-    { 0, 1, 0x161313ff },    // DEFAULT_BASE_COLOR_NORMAL 
-    { 0, 2, 0x38f620ff },    // DEFAULT_TEXT_COLOR_NORMAL 
-    { 0, 3, 0xc3fbc6ff },    // DEFAULT_BORDER_COLOR_FOCUSED 
-    { 0, 4, 0x43bf2eff },    // DEFAULT_BASE_COLOR_FOCUSED 
-    { 0, 5, 0xdcfadcff },    // DEFAULT_TEXT_COLOR_FOCUSED 
-    { 0, 6, 0x1f5b19ff },    // DEFAULT_BORDER_COLOR_PRESSED 
-    { 0, 7, 0x43ff28ff },    // DEFAULT_BASE_COLOR_PRESSED 
-    { 0, 8, 0x1e6f15ff },    // DEFAULT_TEXT_COLOR_PRESSED 
-    { 0, 9, 0x223b22ff },    // DEFAULT_BORDER_COLOR_DISABLED 
-    { 0, 10, 0x182c18ff },    // DEFAULT_BASE_COLOR_DISABLED 
-    { 0, 11, 0x244125ff },    // DEFAULT_TEXT_COLOR_DISABLED 
-    { 0, 16, 0x00000010 },    // DEFAULT_TEXT_SIZE 
-    { 0, 17, 0x00000000 },    // DEFAULT_TEXT_SPACING 
-    { 0, 18, 0xe6fce3ff },    // DEFAULT_LINE_COLOR 
-    { 0, 19, 0x0c1505ff },    // DEFAULT_BACKGROUND_COLOR 
+    { 0, 0, (int)0x1c8d00ff },    // DEFAULT_BORDER_COLOR_NORMAL 
+    { 0, 1, (int)0x161313ff },    // DEFAULT_BASE_COLOR_NORMAL 
+    { 0, 2, (int)0x38f620ff },    // DEFAULT_TEXT_COLOR_NORMAL 
+    { 0, 3, (int)0xc3fbc6ff },    // DEFAULT_BORDER_COLOR_FOCUSED 
+    { 0, 4, (int)0x43bf2eff },    // DEFAULT_BASE_COLOR_FOCUSED 
+    { 0, 5, (int)0xdcfadcff },    // DEFAULT_TEXT_COLOR_FOCUSED 
+    { 0, 6, (int)0x1f5b19ff },    // DEFAULT_BORDER_COLOR_PRESSED 
+    { 0, 7, (int)0x43ff28ff },    // DEFAULT_BASE_COLOR_PRESSED 
+    { 0, 8, (int)0x1e6f15ff },    // DEFAULT_TEXT_COLOR_PRESSED 
+    { 0, 9, (int)0x223b22ff },    // DEFAULT_BORDER_COLOR_DISABLED 
+    { 0, 10, (int)0x182c18ff },    // DEFAULT_BASE_COLOR_DISABLED 
+    { 0, 11, (int)0x244125ff },    // DEFAULT_TEXT_COLOR_DISABLED 
+    { 0, 16, (int)0x00000010 },    // DEFAULT_TEXT_SIZE 
+    { 0, 17, (int)0x00000000 },    // DEFAULT_TEXT_SPACING 
+    { 0, 18, (int)0xe6fce3ff },    // DEFAULT_LINE_COLOR 
+    { 0, 19, (int)0x0c1505ff },    // DEFAULT_BACKGROUND_COLOR 
 };
 
 // WARNING: This style uses a custom font:  (size: 16, spacing: 0)
